+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/ast/ASTExpression.java
@@ -20,4 +20,32 @@ public class ASTExpression extends AbstractJavaTypeNode {
     public Object jjtAccept(JavaParserVisitor visitor, Object data) {
         return visitor.visit(this, data);
     }
+
+    public boolean isStandAlonePrimitive() {
+        if (jjtGetNumChildren() != 1) {
+            return false;
+        }
+
+        ASTPrimaryExpression primaryExpression = getFirstChildOfType(ASTPrimaryExpression.class);
+
+        if (primaryExpression == null || primaryExpression.jjtGetNumChildren() != 1) {
+            return false;
+        }
+
+        ASTPrimaryPrefix primaryPrefix = primaryExpression.getFirstChildOfType(ASTPrimaryPrefix.class);
+
+        if (primaryPrefix == null || primaryPrefix.jjtGetNumChildren() != 1) {
+            return false;
+        }
+
+        ASTLiteral literal = primaryPrefix.getFirstChildOfType(ASTLiteral.class);
+
+        if (literal == null || literal.isStringLiteral()
+                || (literal.jjtGetNumChildren() != 0 && literal.jjtGetChild(0) instanceof ASTNullLiteral)) {
+            return false;
+        }
+
+        // byte, short, char, int, long, float, double, boolean
+        return true;
+    }
 }