+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/typeresolution/typeinference/TypeInferenceResolver.java
@@ -23,12 +23,44 @@ import net.sourceforge.pmd.lang.java.typeresolution.typedefinition.JavaTypeDefin
 
 public final class TypeInferenceResolver {
 
+
     public static class ResolutionFailedException extends RuntimeException {
 
     }
 
     private TypeInferenceResolver() {
 
+    }
+
+    public static List<JavaTypeDefinition> inferTypes(List<Constraint> constraints, List<Bound> bounds,
+                                                      List<Variable> variables) {
+
+        List<Bound> newBounds = new ArrayList<>();
+        while (!constraints.isEmpty()) {
+            List<BoundOrConstraint> reduceResult = constraints.remove(constraints.size() - 1).reduce();
+
+            for (BoundOrConstraint boundOrConstraint : reduceResult) {
+                if (boundOrConstraint instanceof Bound) {
+                    newBounds.add((Bound) boundOrConstraint);
+                } else if (boundOrConstraint instanceof Constraint) {
+                    constraints.add((Constraint) boundOrConstraint);
+                } else {
+                    throw new IllegalStateException("Unknown BoundOrConstraint subclass");
+                }
+            }
+
+            constraints.addAll(incorporateBounds(bounds, newBounds));
+            newBounds.clear();
+        }
+
+        Map<Variable, JavaTypeDefinition> resolutionResult = resolveVariables(bounds);
+        List<JavaTypeDefinition> result = new ArrayList<>();
+
+        for (Variable variable : variables) {
+            result.add(resolutionResult.get(variable));
+        }
+
+        return result;
     }
 
     public static JavaTypeDefinition lub(List<JavaTypeDefinition> types) {
@@ -509,6 +541,8 @@ public final class TypeInferenceResolver {
                 // α = U and S <: T imply ‹S[α:=U] <: T[α:=U]› TODO
             }
         }
+
+        currentBounds.addAll(newBounds);
 
         return newConstraints;
     }